feat: add notifications to posts and comments

+++ src/main/java/com/booklink/backend/service/impl/CommentServiceImpl.java
@@ -7,40 +7,50 @@ import com.booklink.backend.dto.forum.ForumDtoFactory;
 import com.booklink.backend.dto.post.PostDto;
 import com.booklink.backend.exception.*;
 import com.booklink.backend.model.Comment;
+import com.booklink.backend.model.Post;
+import com.booklink.backend.model.User;
 import com.booklink.backend.repository.CommentRepository;
-import com.booklink.backend.service.CommentService;
-import com.booklink.backend.service.PostService;
-import com.booklink.backend.service.ReactionService;
+import com.booklink.backend.service.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
 public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
     private final PostService postService;
+    private final ForumService forumService;
+    private final NotificationService notificationService;
     private final ForumDtoFactory forumDtoFactory;
     private final ReactionService<Comment> reactionService;
 
-    public CommentServiceImpl(CommentRepository commentRepository, PostService postService, ForumDtoFactory forumDtoFactory, ReactionService<Comment> reactionService) {
+    public CommentServiceImpl(CommentRepository commentRepository, PostService postService, ForumService forumService, NotificationService notificationService, ForumDtoFactory forumDtoFactory, ReactionService<Comment> reactionService) {
         this.commentRepository = commentRepository;
         this.postService = postService;
+        this.forumService = forumService;
+        this.notificationService = notificationService;
         this.forumDtoFactory = forumDtoFactory;
         this.reactionService = reactionService;
     }
 
     @Override
     public CommentDto createComment(CreateCommentDto createCommentDto, Long userId) {
-        PostDto postDto = postService.getPostById(createCommentDto.getPostId());
-        boolean isMember = forumDtoFactory.isMember(postDto.getForumId(), userId);
+        Post post = postService.getPostEntity(createCommentDto.getPostId());
+        boolean isMember = forumDtoFactory.isMember(post.getForumId(), userId);
 
         if (!isMember) throw new MemberDoesntBelongForumException("No puedes comentar en este posteo");
 
         Comment commentToSave = Comment.from(createCommentDto, userId);
         Comment savedComment = commentRepository.save(commentToSave);
+
+        List<Long> usersId = post.getComments().stream().map(Comment::getUserId).distinct().collect(Collectors.toList());
+        notificationService.createCommentNotification(userId, post.getUserId(), usersId, post.getId(), savedComment.getId());
+
         return CommentDto.from(savedComment);
     }
 
@@ -55,6 +65,7 @@ public class CommentServiceImpl implements CommentService {
         Optional<Comment> commentOptional = commentRepository.findById(id);
         return commentOptional.orElseThrow(() -> new NotFoundException("El comentario no fue encontrado"));
     }
+
     @Override
     public void deleteComment(Long id, Long userId) {
         Comment comment = getCommentEntityById(id);
@@ -63,7 +74,7 @@ public class CommentServiceImpl implements CommentService {
         boolean isCommentCreator = comment.getUserId().equals(userId);
         boolean isForumOwner = forumDtoFactory.isForumOwner(forumid, userId);
 
-        if (!isCommentCreator && !isForumOwner) throw new UserNotOwnerException("No tienes permiso para eliminar este comentario");
+        if (!isCommentCreator || !isForumOwner) throw new UserNotOwnerException("No tienes permiso para eliminar este comentario");
 
         commentRepository.deleteById(id);
     }
